pyproject.toml
[tool.poetry]
name = "policy-as-code"
version = "0.1.0"
description = "A simple linter for markdown files."
authors = ["Jules <jules@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-frontmatter = "^1.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
